# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Play Later API",
    "description": "A gaming backlog management application",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/playthroughs": {
      "get": {
        "tags": [
          "playthroughs"
        ],
        "summary": "List Playthroughs",
        "operationId": "list_playthroughs_api_v1_playthroughs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaythroughStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by playthrough status",
              "default": null,
              "title": "Status"
            },
            "description": "Filter by playthrough status"
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by platforms",
              "default": null,
              "title": "Platform"
            },
            "description": "Filter by platforms"
          },
          {
            "name": "rating_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum rating (1-10)",
              "default": null,
              "title": "Rating Min"
            },
            "description": "Minimum rating (1-10)"
          },
          {
            "name": "rating_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum rating (1-10)",
              "default": null,
              "title": "Rating Max"
            },
            "description": "Maximum rating (1-10)"
          },
          {
            "name": "play_time_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum play time in hours",
              "default": null,
              "title": "Play Time Min"
            },
            "description": "Minimum play time in hours"
          },
          {
            "name": "play_time_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum play time in hours",
              "default": null,
              "title": "Play Time Max"
            },
            "description": "Maximum play time in hours"
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by difficulty settings",
              "default": null,
              "title": "Difficulty"
            },
            "description": "Filter by difficulty settings"
          },
          {
            "name": "playthrough_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by playthrough type",
              "default": null,
              "title": "Playthrough Type"
            },
            "description": "Filter by playthrough type"
          },
          {
            "name": "started_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Started after date (YYYY-MM-DD)",
              "default": null,
              "title": "Started After"
            },
            "description": "Started after date (YYYY-MM-DD)"
          },
          {
            "name": "started_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Started before date (YYYY-MM-DD)",
              "default": null,
              "title": "Started Before"
            },
            "description": "Started before date (YYYY-MM-DD)"
          },
          {
            "name": "completed_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Completed after date (YYYY-MM-DD)",
              "default": null,
              "title": "Completed After"
            },
            "description": "Completed after date (YYYY-MM-DD)"
          },
          {
            "name": "completed_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Completed before date (YYYY-MM-DD)",
              "default": null,
              "title": "Completed Before"
            },
            "description": "Completed before date (YYYY-MM-DD)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in game titles and notes",
              "default": null,
              "title": "Search"
            },
            "description": "Search in game titles and notes"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PlaythroughSortBy",
              "description": "Sort field",
              "default": "updated_at"
            },
            "description": "Sort field"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Sort direction",
              "default": "desc"
            },
            "description": "Sort direction"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "playthroughs"
        ],
        "summary": "Create Playthrough",
        "operationId": "create_playthrough_api_v1_playthroughs_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaythroughCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playthroughs/backlog": {
      "get": {
        "tags": [
          "playthroughs"
        ],
        "summary": "Get Backlog",
        "operationId": "get_backlog_api_v1_playthroughs_backlog_get",
        "parameters": [
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by collection priority",
              "default": null,
              "title": "Priority"
            },
            "description": "Filter by collection priority"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacklogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playthroughs/playing": {
      "get": {
        "tags": [
          "playthroughs"
        ],
        "summary": "Get Playing",
        "operationId": "get_playing_api_v1_playthroughs_playing_get",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by platform",
              "default": null,
              "title": "Platform"
            },
            "description": "Filter by platform"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playthroughs/completed": {
      "get": {
        "tags": [
          "playthroughs"
        ],
        "summary": "Get Completed",
        "operationId": "get_completed_api_v1_playthroughs_completed_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by completion year",
              "default": null,
              "title": "Year"
            },
            "description": "Filter by completion year"
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by platform",
              "default": null,
              "title": "Platform"
            },
            "description": "Filter by platform"
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by minimum rating",
              "default": null,
              "title": "Min Rating"
            },
            "description": "Filter by minimum rating"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playthroughs/stats": {
      "get": {
        "tags": [
          "playthroughs"
        ],
        "summary": "Get Stats",
        "operationId": "get_stats_api_v1_playthroughs_stats_get",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/playthroughs/bulk": {
      "post": {
        "tags": [
          "playthroughs"
        ],
        "summary": "Bulk Playthrough Operations",
        "operationId": "bulk_playthrough_operations_api_v1_playthroughs_bulk_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "X-User-Id"
            }
          },
          {
            "name": "session_user",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Session User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaythroughBulkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughBulkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlaythroughListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlaythroughListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "filters_applied": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters Applied"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_count",
          "limit",
          "offset"
        ],
        "title": "PlaythroughListResponse"
      },
...
      "PlaythroughSortBy": {
        "type": "string",
        "enum": [
          "title",
          "status",
          "started_at",
          "completed_at",
          "updated_at",
          "rating",
          "play_time_hours",
          "platform"
        ],
        "title": "PlaythroughSortBy"
      },
...
      "PlaythroughStatus": {
        "type": "string",
        "enum": [
          "PLANNING",
          "PLAYING",
          "COMPLETED",
          "DROPPED",
          "ON_HOLD",
          "MASTERED"
        ],
        "title": "PlaythroughStatus"
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder"
      },
...
    }
  }
}